personal-study-assistant/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
openai>=1.0.0
google-generativeai>=0.1.0
python-dotenv>=0.21.0